Part-3/ai-agent-commerce/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
huggingface-hub==0.25.2
pytest==8.3.3
